requirements.txt
docxcompose==1.4.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
python_docx==0.8.11
PyYAML==5.4

